Move methods from Command class to Cache/Cleaner class
