+++ src/Cache/Cleaner.php
@@ -2,8 +2,40 @@
 
 namespace Timber\Cache;
 
+use Timber\Loader;
+
 class Cleaner
 {
+    public static function clear_cache($mode = 'all')
+    {
+        if (is_array($mode)) {
+            $mode = reset($mode);
+        }
+        if ($mode == 'all') {
+            $twig_cache = self::clear_cache_twig();
+            $timber_cache = self::clear_cache_timber();
+            if ($twig_cache && $timber_cache) {
+                return true;
+            }
+        } elseif ($mode == 'twig') {
+            return self::clear_cache_twig();
+        } elseif ($mode == 'timber') {
+            return self::clear_cache_timber();
+        }
+    }
+
+    public static function clear_cache_timber()
+    {
+        $loader = new Loader();
+        return $loader->clear_cache_timber();
+    }
+
+    public static function clear_cache_twig()
+    {
+        $loader = new Loader();
+        return $loader->clear_cache_twig();
+    }
+
     protected static function delete_transients_single_site()
     {
         global $wpdb;
